cloud_functions/requirements.txt
google-cloud-storage
google-cloud-vision
google-cloud-firestore
functions-framework
firebase-admin
python-dotenv
google-genai